Enable system appropriate text AA, a la core LAFs
Use reflection to set up and utilize the antialiasing features hidden in
the sun.* packages.  The reflection is safe, ensuring that failures to 
initialize or utilize the reflected code do not leak.  The fallback for
reflection failures is the current painting approach.

This approach works with Oracle JREs.  Given the new Oracle JRE on Mac,
this code addresses Windows, Mac, and Linux releases effectively.  Note,
this change will not affect pre-Oracle, Apple JREs.